fixed form
menambahkan gambar pada crud

+++ resources/views/kamar/listKamar.blade.php
@@ -2,26 +2,16 @@
 <div class="container-fluid">
     <div class="row">
         <div class="mt-4 p-0">
-            <h2>Data Kamar {{$kos->nama}}</h2>
+            <h2>Data Kamar {{ $kos->nama }}</h2>
             <!-- menu atas  -->
             <div class="d-flex justify-content-between">
                 <div class="">
-                    <a href="/add-kamar/{{ $kos->id }}" class="btn btn-primary"
-                        >Tambah</a
-                    >
+                    <a href="/add-kamar/{{ $kos->id }}" class="btn btn-primary">Tambah</a>
                 </div>
                 <div class="d-flex">
-                    <form
-                        action="/cari-kamar/{{ $kos->id }}"
-                        method="post"
-                        class="d-flex me-1"
-                    >
+                    <form action="/cari-kamar/{{ $kos->id }}" method="post" class="d-flex me-1">
                         @csrf
-                        <input
-                            type="text"
-                            name="cari"
-                            class="form-control me-1"
-                        />
+                        <input type="text" name="cari" class="form-control me-1" />
                         <button class="btn btn-primary">
                             <i class="bi bi-search"></i>
                         </button>
@@ -29,23 +19,23 @@
                 </div>
             </div>
             <!-- flash message insert -->
-            @if(Session::has('insert'))
+            @if (Session::has('insert'))
             <div class="alert alert-success mt-3">
-                {{ Session::get('pesan')}}
+                {{ Session::get('pesan') }}
             </div>
             @endif
 
             <!-- flash message update -->
-            @if(Session::has('update'))
+            @if (Session::has('update'))
             <div class="alert alert-success mt-3">
-                {{ Session::get('pesan')}}
+                {{ Session::get('pesan') }}
             </div>
             @endif
 
             <!-- flash message delete -->
-            @if(Session::has('delete'))
+            @if (Session::has('delete'))
             <div class="alert alert-danger mt-3">
-                {{ Session::get('pesan')}}
+                {{ Session::get('pesan') }}
             </div>
             @endif
         </div>
@@ -57,6 +47,7 @@
                     <td>Fasilitas</td>
                     <td>Harga</td>
                     <td>Status</td>
+                    <td>Gambar</td>
                     <td>Action</td>
                 </tr>
             </thead>
@@ -64,25 +55,24 @@
             <tbody class="">
                 <tr class="table-striped">
                     <td>{{ $loop->iteration }}</td>
-                    <td>{{ $k->nama}}</td>
-                    <td>{{ $k->fasilitas}}</td>
-                    <td>{{ $k->harga}}</td>
-                    <td>{{ $k->status}}</td>
+                    <td>{{ $k->nama }}</td>
+                    <td>{{ $k->fasilitas }}</td>
+                    <td>{{ $k->harga }}</td>
+                    <td>{{ $k->status }}</td>
+                    <td>
+                        @if ($k->gambar == null)
+                        <p>- </p>
+                        @else
+                        <img src="{{ asset('storage/gambar/' . $k->gambar) }}" width="150px" />
+                        @endif
+                    </td>
                     <td style="width: 100px; white-space: nowrap">
                         <!-- <a href="#" class="btn btn-warning">Detail</a> -->
-                        <a
-                            href="/edit-kamar/{{$k->id}}"
-                            class="btn btn-success"
-                        >
-                            <i class="bi bi-pencil-fill"></i
-                        ></a>
-                        <a
-                            href="/delete-kamar/{{$k->id}}"
-                            onclick="return confirm('Hapus data {{ $k->nama }} ?')"
-                            class="btn btn-danger"
-                        >
-                            <i class="bi bi-trash-fill"></i
-                        ></a>
+                        <a href="/edit-kamar/{{ $k->id }}" class="btn btn-success">
+                            <i class="bi bi-pencil-fill"></i></a>
+                        <a href="/delete-kamar/{{ $k->id }}" onclick="return confirm('Hapus data {{ $k->nama }} ?')"
+                            class="btn btn-danger">
+                            <i class="bi bi-trash-fill"></i></a>
                     </td>
                 </tr>
             </tbody>
